Etch a sketch version 1.0
This etch a sketch web app allows users to draw
on a palette with three different colors (RGB).
The user may also use an eraser button. The user
has the option to clear the grid with a confirmation.
The user can change the size of the grid by entering
a value between 10-99.